Publish environments research and new style README.md
