+++ environments/environment-anaconda.yml
@@ -1,0 +1,111 @@
+name: RL
+channels:
+  - anaconda
+  - conda-forge
+  - defaults
+dependencies:
+  - asttokens=2.4.1=pyhd8ed1ab_0
+  - ca-certificates=2024.7.4=h56e8100_0
+  - certifi=2024.7.4=pyhd8ed1ab_0
+  - comm=0.2.2=pyhd8ed1ab_0
+  - debugpy=1.8.2=py39ha51f57c_0
+  - decorator=5.1.1=pyhd8ed1ab_0
+  - exceptiongroup=1.2.2=pyhd8ed1ab_0
+  - executing=2.0.1=pyhd8ed1ab_0
+  - importlib_metadata=8.0.0=hd8ed1ab_0
+  - ipykernel=6.29.5=pyh4bbf305_0
+  - ipython=8.18.1=pyh7428d3b_3
+  - jedi=0.19.1=pyhd8ed1ab_0
+  - jupyter_client=8.6.2=pyhd8ed1ab_0
+  - jupyter_core=5.7.2=py39hcbf5309_0
+  - krb5=1.21.3=hdf4eb48_0
+  - libsodium=1.0.18=h8d14728_1
+  - matplotlib-inline=0.1.7=pyhd8ed1ab_0
+  - menuinst=2.0.0=py39hd77b12b_0
+  - nest-asyncio=1.6.0=pyhd8ed1ab_0
+  - openssl=3.3.1=h2466b09_2
+  - parso=0.8.4=pyhd8ed1ab_0
+  - pickleshare=0.7.5=py_1003
+  - pip=23.3=py39haa95532_0
+  - platformdirs=4.2.2=pyhd8ed1ab_0
+  - prompt-toolkit=3.0.47=pyha770c72_0
+  - psutil=6.0.0=py39ha55e580_0
+  - pure_eval=0.2.3=pyhd8ed1ab_0
+  - python=3.9.18=h1aa4202_0
+  - python_abi=3.9=2_cp39
+  - pywin32=306=py39h99910a6_2
+  - pyzmq=26.0.3=py39h03e5c00_0
+  - setuptools=68.0.0=py39haa95532_0
+  - sqlite=3.41.2=h2bbff1b_0
+  - stack_data=0.6.2=pyhd8ed1ab_0
+  - tornado=6.4.1=py39ha55e580_0
+  - traitlets=5.14.3=pyhd8ed1ab_0
+  - typing_extensions=4.12.2=pyha770c72_0
+  - ucrt=10.0.22621.0=h57928b3_0
+  - vc=14.2=h21ff451_1
+  - vc14_runtime=14.40.33810=ha82c5b3_20
+  - vs2015_runtime=14.40.33810=h3bf8584_20
+  - wcwidth=0.2.13=pyhd8ed1ab_0
+  - wheel=0.41.2=py39haa95532_0
+  - zeromq=4.3.5=he1f189c_4
+  - pip:
+      - absl-py==2.1.0
+      - astunparse==1.6.3
+      - charset-normalizer==3.3.2
+      - chex==0.1.86
+      - click==8.1.7
+      - colorama==0.4.6
+      - contextlib2==21.6.0
+      - cytoolz==0.12.3
+      - dm-env==1.6
+      - dm-meltingpot==2.1.1.dev4
+      - dm-tree==0.1.8
+      - flatbuffers==24.3.25
+      - gast==0.6.0
+      - google-pasta==0.2.0
+      - grpcio==1.64.1
+      - h5py==3.11.0
+      - idna==3.7
+      - immutabledict==4.2.0
+      - importlib-metadata==8.0.0
+      - jax==0.4.30
+      - jaxlib==0.4.30
+      - keras==3.4.1
+      - libclang==18.1.1
+      - markdown==3.6
+      - markdown-it-py==3.0.0
+      - markupsafe==2.1.5
+      - mdurl==0.1.2
+      - ml-collections==0.1.1
+      - ml-dtypes==0.4.0
+      - namex==0.0.8
+      - networkx==3.2.1
+      - numpy==1.26.4
+      - opencv-python==4.10.0.84
+      - opt-einsum==3.3.0
+      - optree==0.12.1
+      - packaging==24.1
+      - pandas==2.2.2
+      - protobuf==4.25.3
+      - pygame==2.6.0
+      - pygments==2.18.0
+      - python-dateutil==2.9.0.post0
+      - pytz==2024.1
+      - pyyaml==6.0.1
+      - reactivex==4.0.4
+      - requests==2.32.3
+      - rich==13.7.1
+      - scipy==1.13.1
+      - six==1.16.0
+      - tensorboard==2.17.0
+      - tensorboard-data-server==0.7.2
+      - tensorflow==2.17.0
+      - tensorflow-intel==2.17.0
+      - tensorflow-io-gcs-filesystem==0.31.0
+      - termcolor==2.4.0
+      - toolz==0.12.1
+      - tzdata==2024.1
+      - urllib3==2.2.2
+      - werkzeug==3.0.3
+      - wrapt==1.16.0
+      - zipp==3.19.2
