+++ environments/swap-pre-meltingpot/environment-swap1.yml
@@ -1,0 +1,97 @@
+name: RL
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - absl-py=2.1.0=py39haa95532_0
+  - astunparse=1.6.3=py_0
+  - blas=1.0=mkl
+  - bottleneck=1.3.7=py39h9128911_0
+  - brotli-python=1.0.9=py39hd77b12b_8
+  - ca-certificates=2024.7.2=haa95532_0
+  - certifi=2024.7.4=py39haa95532_0
+  - chex=0.1.86=pyhd8ed1ab_0
+  - click=8.1.7=py39haa95532_0
+  - colorama=0.4.6=py39haa95532_0
+  - contextlib2=21.6.0=py39haa95532_0
+  - dm-env=1.6=pyhd8ed1ab_0
+  - google-pasta=0.2.0=pyhd3eb1b0_0
+  - h5py=3.11.0=py39hed405ee_0
+  - hdf5=1.12.1=h51c971a_3
+  - icc_rt=2022.1.0=h6049295_2
+  - idna=3.7=py39haa95532_0
+  - immutabledict=4.2.0=pyhd8ed1ab_0
+  - intel-openmp=2023.1.0=h59b6b97_46320
+  - libabseil=20240116.2=cxx17_h5da7b33_0
+  - libprotobuf=4.25.3=hf2fb9eb_0
+  - menuinst=2.1.1=py39h5da7b33_0
+  - mkl=2023.1.0=h6b88ed4_46358
+  - mkl-service=2.4.0=py39h2bbff1b_1
+  - ml-collections=0.1.1=pyhd8ed1ab_0
+  - networkx=3.2.1=py39haa95532_0
+  - numexpr=2.8.7=py39h2cd9be0_0
+  - openssl=3.0.14=h827c3e9_0
+  - opt-einsum=3.3.0=hd8ed1ab_2
+  - opt_einsum=3.3.0=pyhd3eb1b0_1
+  - packaging=24.1=py39haa95532_0
+  - pandas=2.2.2=py39h5da7b33_0
+  - pip=24.0=py39haa95532_0
+  - protobuf=4.25.3=py39h958608f_0
+  - pybind11-abi=5=hd3eb1b0_0
+  - pysocks=1.7.1=py39haa95532_0
+  - python=3.9.19=h1aa4202_1
+  - python-dateutil=2.9.0post0=py39haa95532_2
+  - python-tzdata=2023.3=pyhd3eb1b0_0
+  - pytz=2024.1=py39haa95532_0
+  - pyyaml=6.0.1=py39h2bbff1b_0
+  - reactivex=4.0.4=pyhd8ed1ab_0
+  - requests=2.32.3=py39haa95532_0
+  - scipy=1.13.1=py39h8640f81_0
+  - setuptools=69.5.1=py39haa95532_0
+  - six=1.16.0=pyhd3eb1b0_1
+  - sqlite=3.45.3=h2bbff1b_0
+  - tbb=2021.8.0=h59b6b97_0
+  - urllib3=2.2.2=py39haa95532_0
+  - vc=14.2=h2eaa2aa_4
+  - vs2015_runtime=14.29.30133=h43f2093_4
+  - werkzeug=3.0.3=py39haa95532_0
+  - wheel=0.43.0=py39haa95532_0
+  - win_inet_pton=1.1.0=py39haa95532_0
+  - yaml=0.2.5=he774522_0
+  - zlib=1.2.13=h8cc25b3_1
+  - pip:
+      - charset-normalizer==3.3.2
+      - cytoolz==0.12.3
+      - dm-meltingpot==2.1.1.dev4
+      - dm-tree==0.1.8
+      - flatbuffers==24.3.25
+      - gast==0.6.0
+      - grpcio==1.64.1
+      - importlib-metadata==8.0.0
+      - jax==0.4.30
+      - jaxlib==0.4.30
+      - keras==3.4.1
+      - libclang==18.1.1
+      - markdown==3.6
+      - markdown-it-py==3.0.0
+      - markupsafe==2.1.5
+      - mdurl==0.1.2
+      - ml-dtypes==0.4.0
+      - namex==0.0.8
+      - numpy==1.26.4
+      - opencv-python==4.10.0.84
+      - optree==0.12.1
+      - pygame==2.6.0
+      - pygments==2.18.0
+      - rich==13.7.1
+      - tensorboard==2.17.0
+      - tensorboard-data-server==0.7.2
+      - tensorflow==2.17.0
+      - tensorflow-intel==2.17.0
+      - tensorflow-io-gcs-filesystem==0.31.0
+      - termcolor==2.4.0
+      - toolz==0.12.1
+      - typing-extensions==4.12.2
+      - tzdata==2024.1
+      - wrapt==1.16.0
+      - zipp==3.19.2
