+++ environments/swap-pre-meltingpot/environment-swap2.yml
@@ -1,0 +1,132 @@
+name: RL
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - _tflow_select=2.1.0=gpu
+  - absl-py=2.1.0=py39haa95532_0
+  - aiohttp=3.9.5=py39h2bbff1b_0
+  - aiosignal=1.2.0=pyhd3eb1b0_0
+  - astunparse=1.6.3=py_0
+  - async-timeout=4.0.3=py39haa95532_0
+  - attrs=23.1.0=py39haa95532_0
+  - blas=1.0=mkl
+  - blinker=1.6.2=py39haa95532_0
+  - bottleneck=1.3.7=py39h9128911_0
+  - brotli-python=1.0.9=py39hd77b12b_8
+  - ca-certificates=2024.7.2=haa95532_0
+  - cachetools=5.3.3=py39haa95532_0
+  - certifi=2024.7.4=py39haa95532_0
+  - cffi=1.16.0=py39h2bbff1b_1
+  - chex=0.1.86=pyhd8ed1ab_0
+  - click=8.1.7=py39haa95532_0
+  - colorama=0.4.6=py39haa95532_0
+  - contextlib2=21.6.0=py39haa95532_0
+  - cryptography=41.0.3=py39h3438e0d_0
+  - cudatoolkit=11.3.1=h59b6b97_2
+  - cudnn=8.2.1=cuda11.3_0
+  - dm-env=1.6=pyhd8ed1ab_0
+  - frozenlist=1.4.0=py39h2bbff1b_0
+  - gast=0.4.0=pyhd3eb1b0_0
+  - giflib=5.2.1=h8cc25b3_3
+  - google-auth=2.29.0=py39haa95532_0
+  - google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
+  - google-pasta=0.2.0=pyhd3eb1b0_0
+  - h5py=3.11.0=py39hed405ee_0
+  - hdf5=1.12.1=h51c971a_3
+  - icc_rt=2022.1.0=h6049295_2
+  - icu=58.2=ha925a31_3
+  - idna=3.7=py39haa95532_0
+  - immutabledict=4.2.0=pyhd8ed1ab_0
+  - intel-openmp=2023.1.0=h59b6b97_46320
+  - jpeg=9e=h2bbff1b_1
+  - keras=2.10.0=py39haa95532_0
+  - keras-preprocessing=1.1.2=pyhd3eb1b0_0
+  - libabseil=20240116.2=cxx17_h5da7b33_0
+  - libcurl=8.7.1=h86230a5_0
+  - libpng=1.6.39=h8cc25b3_0
+  - libprotobuf=3.20.3=h23ce68f_0
+  - libssh2=1.10.0=hcd4344a_2
+  - menuinst=2.1.1=py39h5da7b33_0
+  - mkl=2023.1.0=h6b88ed4_46358
+  - mkl-service=2.4.0=py39h2bbff1b_1
+  - ml-collections=0.1.1=pyhd8ed1ab_0
+  - multidict=6.0.4=py39h2bbff1b_0
+  - networkx=3.2.1=py39haa95532_0
+  - numexpr=2.8.7=py39h2cd9be0_0
+  - oauthlib=3.2.2=py39haa95532_0
+  - openssl=1.1.1w=h2bbff1b_0
+  - opt-einsum=3.3.0=hd8ed1ab_2
+  - opt_einsum=3.3.0=pyhd3eb1b0_1
+  - packaging=24.1=py39haa95532_0
+  - pandas=2.2.2=py39h5da7b33_0
+  - pip=24.0=py39haa95532_0
+  - protobuf=3.20.3=py39hd77b12b_0
+  - pyasn1=0.4.8=pyhd3eb1b0_0
+  - pyasn1-modules=0.2.8=py_0
+  - pybind11-abi=5=hd3eb1b0_0
+  - pycparser=2.21=pyhd3eb1b0_0
+  - pyjwt=2.8.0=py39haa95532_0
+  - pyopenssl=23.2.0=py39haa95532_0
+  - pysocks=1.7.1=py39haa95532_0
+  - python=3.9.18=h6244533_0
+  - python-dateutil=2.9.0post0=py39haa95532_2
+  - python-flatbuffers=2.0=pyhd3eb1b0_0
+  - python-tzdata=2023.3=pyhd3eb1b0_0
+  - pytz=2024.1=py39haa95532_0
+  - pyyaml=6.0.1=py39h2bbff1b_0
+  - reactivex=4.0.4=pyhd8ed1ab_0
+  - requests=2.32.3=py39haa95532_0
+  - requests-oauthlib=2.0.0=py39haa95532_0
+  - rsa=4.7.2=pyhd3eb1b0_1
+  - scipy=1.13.1=py39h8640f81_0
+  - setuptools=69.5.1=py39haa95532_0
+  - six=1.16.0=pyhd3eb1b0_1
+  - snappy=1.1.10=h6c2663c_1
+  - sqlite=3.45.3=h2bbff1b_0
+  - tbb=2021.8.0=h59b6b97_0
+  - tensorboard=2.10.0=py39haa95532_0
+  - tensorboard-data-server=0.6.1=py39haa95532_0
+  - tensorboard-plugin-wit=1.8.1=py39haa95532_0
+  - tensorflow=2.10.0=gpu_py39h9bca9fa_0
+  - tensorflow-base=2.10.0=gpu_py39h9761872_0
+  - tensorflow-estimator=2.10.0=py39haa95532_0
+  - typing_extensions=4.11.0=py39haa95532_0
+  - urllib3=2.2.2=py39haa95532_0
+  - vc=14.2=h2eaa2aa_4
+  - vs2015_runtime=14.29.30133=h43f2093_4
+  - werkzeug=3.0.3=py39haa95532_0
+  - wheel=0.43.0=py39haa95532_0
+  - win_inet_pton=1.1.0=py39haa95532_0
+  - yaml=0.2.5=he774522_0
+  - yarl=1.9.3=py39h2bbff1b_0
+  - zlib=1.2.13=h8cc25b3_1
+  - pip:
+      - charset-normalizer==3.3.2
+      - cytoolz==0.12.3
+      - dm-meltingpot==2.1.1.dev4
+      - dm-tree==0.1.8
+      - flatbuffers==24.3.25
+      - grpcio==1.64.1
+      - importlib-metadata==8.0.0
+      - jax==0.4.30
+      - jaxlib==0.4.30
+      - libclang==18.1.1
+      - markdown==3.6
+      - markdown-it-py==3.0.0
+      - markupsafe==2.1.5
+      - mdurl==0.1.2
+      - ml-dtypes==0.4.0
+      - namex==0.0.8
+      - numpy==1.26.4
+      - opencv-python==4.10.0.84
+      - optree==0.12.1
+      - pygame==2.6.0
+      - pygments==2.18.0
+      - rich==13.7.1
+      - termcolor==2.4.0
+      - toolz==0.12.1
+      - typing-extensions==4.12.2
+      - tzdata==2024.1
+      - wrapt==1.16.0
+      - zipp==3.19.2
