+++ environments/swap-pre-meltingpot/environment-swap4.yml
@@ -1,0 +1,167 @@
+name: RL
+channels:
+  - anaconda
+  - conda-forge
+  - defaults
+dependencies:
+  - _tflow_select=2.3.0=mkl
+  - absl-py=2.1.0=pyhd8ed1ab_0
+  - aiohttp=3.9.5=py39ha55989b_0
+  - aiosignal=1.3.1=pyhd8ed1ab_0
+  - astunparse=1.6.3=pyhd8ed1ab_0
+  - async-timeout=4.0.3=pyhd8ed1ab_0
+  - attrs=23.2.0=pyh71513ae_0
+  - blas=2.122=openblas
+  - blas-devel=3.9.0=22_win64_openblas
+  - blinker=1.8.2=pyhd8ed1ab_0
+  - bottleneck=1.4.0=py39h4b0a98a_1
+  - brotli-python=1.1.0=py39h99910a6_1
+  - ca-certificates=2024.7.4=h56e8100_0
+  - cachetools=5.4.0=pyhd8ed1ab_0
+  - certifi=2024.7.4=pyhd8ed1ab_0
+  - cffi=1.16.0=py39ha55989b_0
+  - charset-normalizer=3.3.2=pyhd8ed1ab_0
+  - chex=0.1.86=pyhd8ed1ab_0
+  - click=8.1.7=win_pyh7428d3b_0
+  - colorama=0.4.6=pyhd8ed1ab_0
+  - contextlib2=21.6.0=pyhd8ed1ab_0
+  - cryptography=41.0.3=py39h3438e0d_0
+  - cuda-version=11.8=h70ddcb2_3
+  - cudatoolkit=11.8.0=h09e9e62_13
+  - cudnn=8.9.7.29=he6de189_3
+  - cytoolz=0.12.3=py39ha55989b_0
+  - dm-env=1.6=pyhd8ed1ab_0
+  - dm-tree=0.1.8=py39h018ea92_4
+  - flatbuffers=24.3.25=h63175ca_0
+  - frozenlist=1.4.1=py39ha55989b_0
+  - gast=0.4.0=pyh9f0ad1d_0
+  - giflib=5.2.2=h64bf75a_0
+  - google-auth=2.32.0=pyhff2d567_0
+  - google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
+  - google-pasta=0.2.0=pyh8c360ce_0
+  - h2=4.1.0=pyhd8ed1ab_0
+  - h5py=3.11.0=py39hed405ee_0
+  - hdf5=1.12.1=nompi_h2a0e4a3_104
+  - hpack=4.0.0=pyh9f0ad1d_0
+  - hyperframe=6.0.1=pyhd8ed1ab_0
+  - icc_rt=2022.1.0=h6049295_2
+  - icu=58.2=ha925a31_3
+  - idna=3.7=pyhd8ed1ab_0
+  - immutabledict=4.2.0=pyhd8ed1ab_0
+  - importlib-metadata=8.0.0=pyha770c72_0
+  - intel-openmp=2024.2.0=h57928b3_980
+  - jpeg=9e=hcfcfb64_3
+  - keras=2.10.0=pyhd8ed1ab_0
+  - keras-preprocessing=1.1.2=pyhd8ed1ab_0
+  - libabseil=20240116.2=cxx17_he0c23c2_1
+  - libblas=3.9.0=22_win64_openblas
+  - libcblas=3.9.0=22_win64_openblas
+  - libclang=18.1.7=default_h204ce9a_0
+  - libclang13=18.1.7=default_h97ce8ae_0
+  - libcurl=8.7.1=h86230a5_0
+  - libflang=5.0.0=h6538335_20180525
+  - libhwloc=2.11.1=default_h8125262_1000
+  - libiconv=1.17=hcfcfb64_2
+  - liblapack=3.9.0=22_win64_openblas
+  - liblapacke=3.9.0=22_win64_openblas
+  - libopenblas=0.3.27=pthreads_hc140b1d_0
+  - libpng=1.6.43=h19919ed_0
+  - libprotobuf=3.20.3=h12be248_0
+  - libsqlite=3.46.0=h2466b09_0
+  - libssh2=1.10.0=h680486a_3
+  - libxml2=2.12.7=h283a6d9_1
+  - libzlib=1.2.13=h2466b09_6
+  - libzlib-wapi=1.2.13=h2466b09_6
+  - llvm-meta=5.0.0=0
+  - m2w64-gcc-libgfortran=5.3.0=6
+  - m2w64-gcc-libs=5.3.0=7
+  - m2w64-gcc-libs-core=5.3.0=7
+  - m2w64-gmp=6.1.0=2
+  - m2w64-libwinpthread-git=5.0.0.4634.697f757=2
+  - markdown=3.6=pyhd8ed1ab_0
+  - markdown-it-py=3.0.0=pyhd8ed1ab_0
+  - markupsafe=2.1.5=py39ha55989b_0
+  - mdurl=0.1.2=pyhd8ed1ab_0
+  - menuinst=2.1.1=py39ha51f57c_0
+  - mkl=2023.2.0=h6a75c08_49573
+  - mkl-service=2.4.1=py39hd3ea79f_0
+  - ml-collections=0.1.1=pyhd8ed1ab_0
+  - msys2-conda-epoch=20160418=1
+  - multidict=6.0.5=py39ha55989b_0
+  - namex=0.0.8=pyhd8ed1ab_0
+  - networkx=3.2.1=pyhd8ed1ab_0
+  - numexpr=2.7.3=py39h2e25243_2
+  - oauthlib=3.2.2=pyhd8ed1ab_0
+  - openblas=0.3.27=pthreads_h3721920_0
+  - openmp=5.0.0=vc14_1
+  - openssl=1.1.1w=hcfcfb64_0
+  - opt-einsum=3.3.0=hd8ed1ab_2
+  - opt_einsum=3.3.0=pyhc1e730c_2
+  - optree=0.12.1=py39h2b77a98_0
+  - packaging=24.1=pyhd8ed1ab_0
+  - pandas=2.2.2=py39h2366fc2_1
+  - pip=24.0=pyhd8ed1ab_0
+  - protobuf=3.20.3=py39hcbf5309_1
+  - pthreads-win32=2.9.1=hfa6e2cd_3
+  - pyasn1=0.6.0=pyhd8ed1ab_0
+  - pyasn1-modules=0.4.0=pyhd8ed1ab_0
+  - pybind11-abi=5=hd3eb1b0_0
+  - pycparser=2.22=pyhd8ed1ab_0
+  - pygments=2.18.0=pyhd8ed1ab_0
+  - pyjwt=2.8.0=pyhd8ed1ab_1
+  - pyopenssl=23.2.0=pyhd8ed1ab_1
+  - pysocks=1.7.1=pyh0701188_6
+  - python=3.9.18=h6244533_0
+  - python-dateutil=2.9.0post0=py39haa95532_2
+  - python-flatbuffers=24.3.25=pyh59ac667_0
+  - python-tzdata=2024.1=pyhd8ed1ab_0
+  - python_abi=3.9=2_cp39
+  - pytz=2024.1=pyhd8ed1ab_0
+  - pyu2f=0.1.5=pyhd8ed1ab_0
+  - pyyaml=6.0.1=py39ha55989b_1
+  - reactivex=4.0.4=pyhd8ed1ab_0
+  - requests=2.32.3=pyhd8ed1ab_0
+  - requests-oauthlib=2.0.0=pyhd8ed1ab_0
+  - rich=13.7.1=pyhd8ed1ab_0
+  - rsa=4.9=pyhd8ed1ab_0
+  - scipy=1.13.1=py39h1a10956_0
+  - setuptools=71.0.4=pyhd8ed1ab_0
+  - six=1.16.0=pyh6c4a22f_0
+  - snappy=1.2.1=h23299a8_0
+  - sqlite=3.46.0=h2466b09_0
+  - tbb=2021.12.0=hc790b64_3
+  - tensorboard=2.10.1=pyhd8ed1ab_0
+  - tensorboard-data-server=0.6.1=py39hcbf5309_4
+  - tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
+  - tensorflow=2.10.0=mkl_py39ha510bab_0
+  - tensorflow-base=2.10.0=mkl_py39h6a7f48e_0
+  - tensorflow-estimator=2.10.0=py39haa95532_0
+  - termcolor=2.4.0=pyhd8ed1ab_0
+  - toolz=0.12.1=pyhd8ed1ab_0
+  - typing-extensions=4.12.2=hd8ed1ab_0
+  - typing_extensions=4.12.2=pyha770c72_0
+  - tzdata=2023c=h04d1e81_0
+  - ucrt=10.0.22621.0=h57928b3_0
+  - urllib3=2.2.2=pyhd8ed1ab_1
+  - vc=14.3=h8a93ad2_20
+  - vc14_runtime=14.40.33810=ha82c5b3_20
+  - vs2015_runtime=14.40.33810=h3bf8584_20
+  - werkzeug=3.0.3=pyhd8ed1ab_0
+  - wheel=0.43.0=pyhd8ed1ab_1
+  - win_inet_pton=1.1.0=pyhd8ed1ab_6
+  - wrapt=1.16.0=py39ha55989b_0
+  - yaml=0.2.5=h8ffe710_2
+  - yarl=1.9.4=py39ha55989b_0
+  - zipp=3.19.2=pyhd8ed1ab_0
+  - zlib=1.2.13=h2466b09_6
+  - zstandard=0.23.0=py39h9bf74da_0
+  - zstd=1.5.6=h0ea2cb4_0
+  - pip:
+      - dm-meltingpot==2.1.1.dev4
+      - grpcio==1.64.1
+      - jax==0.4.30
+      - jaxlib==0.4.30
+      - ml-dtypes==0.4.0
+      - numpy==1.26.4
+      - opencv-python==4.10.0.84
+      - pygame==2.6.0
